Extracted interface from MailBackend to prevent mail from tests
